+++ mapbox_baselayer/admin.py
@@ -10,5 +10,7 @@ class TileInLine(admin.StackedInline):
 
 @admin.register(models.MapBaseLayer)
 class MapBaseLayerAdmin(admin.ModelAdmin):
-    list_display = ('name', 'slug', 'order')
+    list_display = ('name', 'slug', 'is_overlay', 'order')
+    list_filter = ('is_overlay',)
+    search_fields = ('name', 'slug')
     inlines = [TileInLine, ]